float double option

+++ project/test.glsl
@@ -5,7 +5,7 @@
 layout(local_size_x = 1, local_size_y = 2, local_size_z = 1) in;
 
 layout(binding = 0, set = 0) buffer Test{
-    float test_float;
+    double test_float;
 };
 
 layout(binding = 1, set = 0) buffer Boo{
